Fix alignment issues in hero and second section
The main text area in hero was given a max-width so that it wouldn't
be able to increase past a point in order to retain the look that
the model assignment in cases where the window size is enlarged. Zombie
code from this section was also removed.

The text below the images in section two was also given a max width
so that it better suits the width of the image on top of it.